refactor(cli): pull out logic to fetch from upstream
