+++ packages/cli/src/commands/release.js
@@ -11,6 +11,7 @@ const parse = require('@commitlint/parse');
 const execa = require('execa');
 const { prompt } = require('inquirer');
 const semver = require('semver');
+const { fetchLatestFromUpstream } = require('../git');
 const { createLogger, displayBanner } = require('../logger');
 
 // All supported commit types from our conventional-changelog preset
@@ -126,31 +127,6 @@ async function release({ bump }) {
   await execa('git', ['commit', '-m', versionCommitMessage]);
 
   logger.stop();
-}
-
-/**
- * For certain release types, we want to be certain that our base branch is
- * up-to-date with the upstream remote. This helper will first check that the
- * upstream remote exists, and create it if it does not, and then will pull the
- * latest changes into the local project.
- *
- * @returns {void}
- */
-async function fetchLatestFromUpstream() {
-  try {
-    // This command will fail is no upstream is present, with `catch` we can
-    // create the appropriate remote before running the next commands
-    await execa('git', ['remote', 'get-url', 'upstream']);
-  } catch {
-    await execa('git', [
-      'remote',
-      'add',
-      'upstream',
-      'git@github.com:carbon-design-system/carbon.git',
-    ]);
-  }
-  await execa('git', ['checkout', 'master']);
-  await execa('git', ['pull', 'upstream', 'master']);
 }
 
 /**
